[ new ] support for ranges 'any of' & 'any but'
Ranges adequately generalise the notions of the empty language (which
is "any of" the empty list) and the language of all the one letter long
words (which is "any but" of the empty list).

Adding them was a piece of cake. The function parsing regexp is even
scarier than yesterday. Oh well...